+++ changeo/Commandline.py
@@ -19,6 +19,24 @@ from presto.IO import printWarning, printError
 from changeo.Defaults import choices_format, default_format
 
 
+def yamlDict(file):
+    """
+    Returns a dictionary from a yaml file
+
+    Arguments:
+      file (str): simple yaml file with rows in the form 'argument: value'.
+
+    Returns:
+      dict: dictionary of key:value pairs in the file.
+    """
+    try:
+        yaml_dict = dict(yaml.load(open(file, 'r')))
+    except:
+        printError('YAML file is invalid.')
+
+    return yaml_dict
+
+
 def yamlArguments(file, args):
     """
     Returns a dictionary of argument from a yaml file
@@ -30,11 +48,8 @@ def yamlArguments(file, args):
     Returns:
       dict: dictionary of arguments.
     """
-    try:
-        yaml_args = yaml.load(open(file, 'r'))
-        yaml_args = {k.lower(): v for k, v in yaml_args.items() if k.lower() in args}
-    except:
-        printError('YAML arguments file is invalid.')
+    yaml_args = yamlDict(file)
+    yaml_args = {k.lower(): v for k, v in yaml_args.items() if k.lower() in args}
 
     return yaml_args